+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - open_log:  Fixed problem with mutable default arguments issue.
 - find_marker:  Fixed problem with mutable default arguments issue.
 - update_marker:  Fixed problem with mutable default arguments issue.
+- get_ignore_msgs:  Fixed problem with mutable default arguments issue.
 
  
 ## [2.1.2] - 2019-01-22