+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - log_2_output:  Fixed problem with mutable default arguments issue.
 - fetch_log:  Fixed problem with mutable default arguments issue.
 - find_marker_array:  Fixed problem with mutable default arguments issue.
+- fetch_log_stdin:  Fixed problem with mutable default arguments issue.
 
  
 ## [2.1.2] - 2019-01-22