+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - fetch_log:  Fixed problem with mutable default arguments issue.
 - find_marker_array:  Fixed problem with mutable default arguments issue.
 - fetch_log_stdin:  Fixed problem with mutable default arguments issue.
+- get_filter_data:  Fixed problem with mutable default arguments issue.
 
  
 ## [2.1.2] - 2019-01-22